fixes USB.get_device_list
Ignore-this: c8f164719d4c5159042628048bf439fe

The comparison "cnt < 0" did not returned the right result because
(size_t) values are unsigned

darcs-hash:20091105091732-c41ad-db8232d9cbd8a9a0a71ce82b897fa22ee05c9e72

+++ src/usb_stubs.c
@@ -183,7 +183,7 @@ CAMLprim value ml_usb_get_device_list(value unit)
   libusb_device **devices;
 
   size_t cnt = libusb_get_device_list(NULL, &devices);
-  if (cnt < 0)
+  if ((int)cnt < 0)
     ml_usb_error(cnt, "get_device_list");
 
   /* Convert the array to a caml list */
